Update ProfileTool.js
Using destructuring assignment to directly assign the width and height values to clientWidth and clientHeight variables respectively, instead of creating new variables for width and height.
Using const instead of let where variables are not reassigned.
Using a for-of loop instead of a for-in loop, as it is more appropriate for arrays and iterable objects.
Removing unnecessary comments.
Removing unnecessary new keyword when calling Utils.projectedRadius
Removing unnecessary new keyword when calling sphere.getWorldPosition