Update ProfileTool.js
Use arrow functions instead of bind() to avoid creating new function objects when the constructor is called. This can improve performance, especially if the constructor is called frequently.

Use a Set instead of an array to keep track of profiles, as Set has faster lookups.

Use a single event listener for both 'profile_added' and 'profile_removed' events and check the type of event to decide whether to call onAdd or onRemove.