Update ProfileTool.js
Move the initialization of the domElement variable to the top of the function, so it is only created once.
Use a ternary operator to assign the value of args.name to profile.name if it exists, otherwise use 'Profile'.
Use let instead of var so that the variable scope is limited to the function.
Remove unnecessary comments.